requirements.txt
mega.py~=1.0.8
anonfile~=0.2.7